friendlyWar/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-environ==0.12.0
django-eventstream==5.3.2
django-grip==3.5.1
django-rest-knox==5.0.2
djangorestframework==3.15.2
environ==1.0
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
MarkupSafe==3.0.2
pillow==11.1.0
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zope.interface==7.2
